// src/entities/Procedure/model/consts/procedureConsts.ts
export enum ProcedureSortField {
    VIEWS = 'views',
    TITLE = 'title',
    CREATED = 'createdAt',
}

export enum ProcedureBlockType {
    INJECTION = 'injectionZones',
    PREPARATION = 'preparations',
    CONTRAINDICATION = 'contraindications',
    REHABILITATION = 'rehabilitations',
    RESULT = 'results',
    PRICE = 'price',
}

// export enum ProcedureType {
//     ALL = 'ALL',
//     IT = 'IT',
//     SCIENCE = 'SCIENCE',
//     ECONOMICS = 'ECONOMICS',
// }

// export enum ProcedureView {
//     BIG = 'BIG',
//     SMALL = 'SMALL',
// }